fix(aside): csv logic
