+++ package.json
@@ -28,6 +28,7 @@
     "autoprefixer": "^10.4.15",
     "eslint": "8.49.0",
     "eslint-config-next": "13.4.19",
+    "flowbite-react": "^0.7.2",
     "postcss": "^8.4.30",
     "tailwindcss": "^3.3.3",
     "typescript": "5.3.3"
