+++ src/components/LoadAside.tsx
@@ -1,0 +1,187 @@
+"use client";
+import Image from "next/image";
+import React, { ChangeEvent, DragEvent, useState } from "react";
+import { Card } from ".";
+import { AddStockCard, Button } from "@/common";
+import Papa from "papaparse";
+import { useStore } from "@/models";
+import axios from "axios";
+
+const EMPTY_FILE_INFO: CsvInfo = {
+  title: "",
+  file: "",
+  currentDate: "",
+} as const;
+
+type CsvInfo = {
+  title: string;
+  file: string;
+  currentDate: string;
+};
+
+const CSVUrls = {
+  MyTeam: "/api/my-team",
+  MyStock: "/api/my-stock",
+} as const;
+
+export const LoadStock = function () {
+  const [csvInfo, setCsvInfo] = useState(EMPTY_FILE_INFO);
+  const [csvFile, setCsvFile] = useState<File | null>(null);
+  const [isLoading, setIsLoading] = useState(false);
+  const [uploadProgress, setUploadProgress] = useState(0);
+
+  const {
+    aside: { csvContext },
+  } = useStore();
+
+  console.log("LoadStock type:", CSVUrls[csvContext]);
+
+  const handleDeleteCard = () => {
+    setCsvInfo(EMPTY_FILE_INFO);
+  };
+
+  const { title, file, currentDate } = csvInfo;
+
+  ///////////
+
+  const postCsvToDatabaseMock = async (parsedData: unknown) => {
+    setIsLoading(true);
+  
+    const totalDuration = 2000; // Duración total de la simulación
+    const updateInterval = 100; // Intervalo de actualización
+  
+    for (let elapsed = 0; elapsed <= totalDuration; elapsed += updateInterval) {
+      await new Promise(resolve => setTimeout(resolve, updateInterval));
+      const progress = Math.round((elapsed / totalDuration) * 100);
+      setUploadProgress(progress);
+    }
+  
+    // Esperar un poco antes de restablecer para ver el progreso completo
+    await new Promise(resolve => setTimeout(resolve, 500));
+    setIsLoading(false);
+    setUploadProgress(0);
+  };
+
+  
+  ///////////
+
+  // const postCsvToDatabase = async (parsedData: unknown) => {
+  //   try {
+  //     setIsLoading(true);
+  //     const apiUrl = CSVUrls[csvContext];
+  //     //post logic here
+
+  //     const config = {
+  //       onUploadProgress: progressEvent => {
+  //         const percentCompleted = Math.round((progressEvent.loaded * 100) / progressEvent.total);
+  //         setUploadProgress(percentCompleted);
+  //       }
+  //     };
+
+  //     //await axios.post(apiUrl, parsedData, config);
+  //     setIsLoading(false);
+  //     setUploadProgress(0);
+  //   } catch (error) {
+  //     setIsLoading(false);
+  //     console.error("Error ->", error);
+  //     setUploadProgress(0);
+  //   }
+  // };
+
+  const onFileChangeHandler = (csvFile: File) => {
+    setCsvFile(csvFile);
+
+    Papa.parse(csvFile, {
+      skipEmptyLines: true, //Delete empty fields
+      header: true, //Parse CSV to JSON
+      complete: function (results) {
+        const { name, size } = csvFile;
+        setCsvInfo({
+          title: name,
+          file: `${(size / 1024).toFixed(2)}kb`,
+          currentDate: new Date().toLocaleString(),
+        });
+        ///console.log("Data:", results.data); //<- log for check data file
+        //postCsvToDatabase(results.data);
+        postCsvToDatabaseMock(results.data);
+      },
+    });
+  };
+
+  // const handleAttachFileClick = () => {
+  //   csvFile ? postCsvToDatabase(csvFile) : console.log("No CSV file selected.");
+  // };
+
+  const handleDrop = (event: DragEvent<HTMLDivElement>) => {
+    event.preventDefault();
+    if (event.dataTransfer.files.length > 0) {
+      const csvFile = event.dataTransfer.files[0];
+      onFileChangeHandler(csvFile);
+    }
+  };
+
+  const handleDragOver = (event: DragEvent<HTMLDivElement>) => {
+    event.preventDefault();
+  };
+
+  const handleFileSelect = (event: ChangeEvent<HTMLInputElement>) => {
+    if (event.target.files && event.target.files.length > 0) {
+      const csvFile = event.target.files[0];
+      onFileChangeHandler(csvFile);
+    }
+  };
+
+  return (
+    <div onDrop={handleDrop} onDragOver={handleDragOver} className="drop-area">
+      <div className="flex flex-col gap-6">
+        <Card className="border-dashed border h-60 flex flex-col gap-4 justify-center">
+          <Image
+            alt="folder icon"
+            src="/svg/folder.svg"
+            width={84}
+            height={88}
+            className="mt-[-20px]"
+          />
+          <p>Drag and drop your CSV file here or</p>
+          <label
+            htmlFor="csvFileSelector"
+            className="cursor-pointer text-blue-500"
+          >
+            <h2 className="text-blue font-lg font-bold font-sans">
+              Select a File
+            </h2>
+          </label>
+          <input
+            type="file"
+            id="csvFileSelector"
+            accept=".csv"
+            className="hidden"
+            onChange={handleFileSelect}
+          />
+        </Card>
+        {csvInfo.title && (
+          <AddStockCard
+            title={title}
+            file={file}
+            currentDate={currentDate}
+            onDeleteClick={handleDeleteCard}
+            isLoading={isLoading}
+            uploadProgress={uploadProgress}
+          />
+        )}
+      </div>
+
+      <div className="fixed bottom-5 w-[85%] flex">
+        <Button
+          variant="primary"
+          body="Attach File"
+          size="big"
+          className="p-3 rounded-md w-full"
+          onClick={() => {
+           // handleAttachFileClick();
+          }}
+        />
+      </div>
+    </div>
+  );
+};
