+++ src/models/aside.store.ts
@@ -5,9 +5,12 @@ type Maybe<T> = T | undefined;
 export const AsideStore = types
   .model({
     type: types.maybe(types.enumeration(ASIDE_TYPES)),
+    csvContext: types.maybe(types.string),
+
   })
   .actions((store) => ({
-    setAside(type: Maybe<AsideType>) {
+    setAside(type: Maybe<AsideType>, csvContext?: string) {
       store.type = type;
+      store.csvContext = csvContext;
     },
   }));
